// packages/state/clients/cw-proposal-multiple.ts
export type Addr = string
export type Uint128 = string
export type Duration =
  | {
      height: number
    }
  | {
      time: number
    }
export type VotingStrategy = {
  single_choice: {
    quorum: PercentageThreshold
    [k: string]: unknown
  }
}
export type PercentageThreshold =
  | {
      majority: {
        [k: string]: unknown
      }
    }
  | {
      percent: Decimal
    }
export type Decimal = string
export interface ConfigResponse {
  dao: Addr
  deposit_info?: CheckedDepositInfo | null
  max_voting_period: Duration
  min_voting_period?: Duration | null
  only_members_execute: boolean
  voting_strategy: VotingStrategy
  [k: string]: unknown
}
export interface CheckedDepositInfo {
  deposit: Uint128
  refund_failed_proposals: boolean
  token: Addr
  [k: string]: unknown
}
// ...
export interface MultipleChoiceVote {
  option_id: number
  [k: string]: unknown
}
export interface DepositInfo {
  deposit: Uint128
  refund_failed_proposals: boolean
  token: DepositToken
  [k: string]: unknown
}
export interface GetVoteResponse {
  vote?: VoteInfo | null
  [k: string]: unknown
}
export interface VoteInfo {
  power: Uint128
  vote: MultipleChoiceVote
  voter: Addr
  [k: string]: unknown
}
export type GovernanceModulesResponse = Addr[]
export interface InfoResponse {
  info: ContractVersion
  [k: string]: unknown
}
export interface ContractVersion {
  contract: string
  version: string
  [k: string]: unknown
}
export interface InstantiateMsg {
  deposit_info?: DepositInfo | null
  max_voting_period: Duration
  min_voting_period?: Duration | null
  only_members_execute: boolean
  voting_strategy: VotingStrategy
  [k: string]: unknown
}
export type MultipleChoiceOptionType = 'None' | 'Standard'
export type Expiration =
  | {
      at_height: number
    }
  | {
      at_time: Timestamp
    }
  | {
      never: {
        [k: string]: unknown
      }
    }
export type Status = 'open' | 'rejected' | 'passed' | 'executed' | 'closed'
export interface ListProposalsResponse {
  proposals: ProposalResponse[]
  [k: string]: unknown
}
export interface ProposalResponse {
  id: number
  proposal: MultipleChoiceProposal
  [k: string]: unknown
}
export interface MultipleChoiceProposal {
  choices: CheckedMultipleChoiceOption[]
  deposit_info?: CheckedDepositInfo | null
  description: string
  expiration: Expiration
  min_voting_period?: Expiration | null
  proposer: Addr
  start_height: number
  status: Status
  title: string
  total_power: Uint128
  votes: MultipleChoiceVotes
  voting_strategy: VotingStrategy
  [k: string]: unknown
}
export interface CheckedMultipleChoiceOption {
  description: string
  index: number
  msgs?: CosmosMsgForEmpty[] | null
  option_type: MultipleChoiceOptionType
  vote_count: Uint128
  [k: string]: unknown
}
export interface MultipleChoiceVotes {
  vote_weights: Uint128[]
  [k: string]: unknown
}
export interface ListVotesResponse {
  votes: VoteInfo[]
  [k: string]: unknown
}
export interface MigrateMsg {
  [k: string]: unknown
}
export type ProposalCountResponse = number
export interface ProposalHooksResponse {
  hooks: string[]
  [k: string]: unknown
}
// ...
export interface ReverseProposalsResponse {
  proposals: ProposalResponse[]
  [k: string]: unknown
}
export interface VoteHooksResponse {
  hooks: string[]
  [k: string]: unknown
}
export interface VoteResponse {
  vote?: VoteInfo | null
  [k: string]: unknown
}
export interface CwProposalMultipleReadOnlyInterface {
  contractAddress: string
  config: () => Promise<ConfigResponse>
  proposal: ({
    proposalId,
  }: {
    proposalId: number
  }) => Promise<ProposalResponse>
  listProposals: ({
    limit,
    startAfter,
  }: {
    limit?: number
    startAfter?: number
  }) => Promise<ListProposalsResponse>
  reverseProposals: ({
    limit,
    startBefore,
  }: {
    limit?: number
    startBefore?: number
  }) => Promise<ReverseProposalsResponse>
  proposalCount: () => Promise<ProposalCountResponse>
  getVote: ({
    proposalId,
    voter,
  }: {
    proposalId: number
    voter: string
  }) => Promise<GetVoteResponse>
  listVotes: ({
    limit,
    proposalId,
    startAfter,
  }: {
    limit?: number
    proposalId: number
    startAfter?: string
  }) => Promise<ListVotesResponse>
  proposalHooks: () => Promise<ProposalHooksResponse>
  voteHooks: () => Promise<VoteHooksResponse>
  info: () => Promise<InfoResponse>
}
export class CwProposalMultipleQueryClient
  implements CwProposalMultipleReadOnlyInterface
{
  client: CosmWasmClient
  contractAddress: string

  constructor(client: CosmWasmClient, contractAddress: string) {
    this.client = client
    this.contractAddress = contractAddress
    this.config = this.config.bind(this)
    this.proposal = this.proposal.bind(this)
    this.listProposals = this.listProposals.bind(this)
    this.reverseProposals = this.reverseProposals.bind(this)
    this.proposalCount = this.proposalCount.bind(this)
    this.getVote = this.getVote.bind(this)
    this.listVotes = this.listVotes.bind(this)
    this.proposalHooks = this.proposalHooks.bind(this)
    this.voteHooks = this.voteHooks.bind(this)
    this.info = this.info.bind(this)
  }

  config = async (): Promise<ConfigResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      config: {},
    })
  }
  proposal = async ({
    proposalId,
  }: {
    proposalId: number
  }): Promise<ProposalResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      proposal: {
        proposal_id: proposalId,
      },
    })
  }
  listProposals = async ({
    limit,
    startAfter,
  }: {
    limit?: number
    startAfter?: number
  }): Promise<ListProposalsResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      list_proposals: {
        limit,
        start_after: startAfter,
      },
    })
  }
  reverseProposals = async ({
    limit,
    startBefore,
  }: {
    limit?: number
    startBefore?: number
  }): Promise<ReverseProposalsResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      reverse_proposals: {
        limit,
        start_before: startBefore,
      },
    })
  }
  proposalCount = async (): Promise<ProposalCountResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      proposal_count: {},
    })
  }
  getVote = async ({
    proposalId,
    voter,
  }: {
    proposalId: number
    voter: string
  }): Promise<GetVoteResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      get_vote: {
        proposal_id: proposalId,
        voter,
      },
    })
  }
  listVotes = async ({
    limit,
    proposalId,
    startAfter,
  }: {
    limit?: number
    proposalId: number
    startAfter?: string
  }): Promise<ListVotesResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      list_votes: {
        limit,
        proposal_id: proposalId,
        start_after: startAfter,
      },
    })
  }
  proposalHooks = async (): Promise<ProposalHooksResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      proposal_hooks: {},
    })
  }
  voteHooks = async (): Promise<VoteHooksResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      vote_hooks: {},
    })
  }
  info = async (): Promise<InfoResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      info: {},
    })
  }
}